+++ test/test_functions.sh
@@ -33,8 +33,13 @@ function testGetRepositoryConfig {
 }
 
 function testGetTargetData {
-	assertEqual "${TESTCLASS}/${FUNCNAME[0]}/1" "$(getTargetData record_count oaipage)" "1"
+	assertEqual "${TESTCLASS}/${FUNCNAME[0]}/1" "$(getTargetData record_count oaipage)" "3"
 	assertEqual "${TESTCLASS}/${FUNCNAME[0]}/2" "$(getTargetData responsedate oaipage)" "2022-02-12T06:07:40Z"
+	assertEqual "${TESTCLASS}/${FUNCNAME[0]}/3" "$(getTargetData format oaipage 1)" "xml"
+	assertEqual "${TESTCLASS}/${FUNCNAME[0]}/4" "$(getTargetData format oaipage 2)" "text"
+	assertEqual "${TESTCLASS}/${FUNCNAME[0]}/5" "$(getTargetData datestamp oaipage 2)" "2022-02-10T21:19:09Z"
+	assertEqual "${TESTCLASS}/${FUNCNAME[0]}/6" "$(getTargetData headerstatus oaipage 3)" "deleted"
+	assertEqual "${TESTCLASS}/${FUNCNAME[0]}/7" "$(getTargetData headerstatus oaipage 2)" ""
 }
 
 function testGetProcessTimeMac {